Applying once again bugfixes
Applying 5990a90fbd5db3112b1457c5f07a66dfb9e64e85 once again.
Bug was introduced again in 03177625c7d216c477cf818690bc0a1b0f4bc8fe and
merged by a mistake.

+++ mt_rng.cpp
@@ -54,7 +54,7 @@ uint32_t get_next(uint32_t min, uint32_t max)
     {
         return min;
     }
-    uint32_t range = (uint32_t)(max - min + 1);
+    uint64_t range = (uint64_t)(max - min) + 1;
     uint32_t random_value = mersenne_twister();
     uint32_t scaled_random = random_value % range;
     return min + scaled_random;
@@ -82,3 +82,4 @@ uint32_t get_normal(double mean, double standard_deviation)
 
     return static_cast<uint32_t>(z0 * standard_deviation + mean);
 }
+
